requirements.txt
asgiref==3.7.2
astroid==2.15.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.2
defusedxml==0.7.1
dill==0.3.7
Django==4.2.3
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pep8==1.7.1
platformdirs==3.9.1
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
pylint==2.17.4
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomlkit==0.11.8
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
